restructure tests directory
