+++ cstar/tests/integration_tests/blueprints/cstar_blueprint_with_netcdf_datasets_template.yaml
