+++ cstar/tests/integration_tests/blueprints/cstar_blueprint_with_yaml_datasets_template.yaml
