+++ cstar/tests/integration_tests/config.py
@@ -35,7 +35,7 @@ TEST_CONFIG = {
     # Remote cases:
     # NetCDF
     "test_case_remote_with_netcdf_datasets": {
-        "template_blueprint_path": f"{TEST_DIRECTORY}/blueprints/cstar_blueprint_with_netcdf_datasets_template.yaml",
+        "template_blueprint_path": f"{TEST_DIRECTORY}/integration_tests/blueprints/cstar_blueprint_with_netcdf_datasets_template.yaml",
         "strs_to_replace": {
             "<input_datasets_location>": "https://github.com/CWorthy-ocean/cstar_blueprint_test_case/raw/main/input_datasets/ROMS",
             "<additional_code_location>": "https://github.com/CWorthy-ocean/cstar_blueprint_test_case.git",
@@ -43,7 +43,7 @@ TEST_CONFIG = {
     },
     # YAML
     "test_case_remote_with_yaml_datasets": {
-        "template_blueprint_path": f"{TEST_DIRECTORY}/blueprints/cstar_blueprint_with_yaml_datasets_template.yaml",
+        "template_blueprint_path": f"{TEST_DIRECTORY}/integration_tests/blueprints/cstar_blueprint_with_yaml_datasets_template.yaml",
         "strs_to_replace": {
             "<input_datasets_location>": "https://github.com/CWorthy-ocean/cstar_blueprint_test_case/raw/main/roms_tools_yaml_files",
             "<additional_code_location>": "https://github.com/CWorthy-ocean/cstar_blueprint_test_case.git",
@@ -52,7 +52,7 @@ TEST_CONFIG = {
     # Local cases:
     # NetCDF
     "test_case_local_with_netcdf_datasets": {
-        "template_blueprint_path": f"{TEST_DIRECTORY}/blueprints/cstar_blueprint_with_netcdf_datasets_template.yaml",
+        "template_blueprint_path": f"{TEST_DIRECTORY}/integration_tests/blueprints/cstar_blueprint_with_netcdf_datasets_template.yaml",
         "strs_to_replace": {
             "<input_datasets_location>": f"{CSTAR_TEST_DATA_DIRECTORY/'input_datasets/ROMS'}",
             "<additional_code_location>": f"{CSTAR_TEST_DATA_DIRECTORY}",
@@ -60,7 +60,7 @@ TEST_CONFIG = {
     },
     # YAML
     "test_case_local_with_yaml_datasets": {
-        "template_blueprint_path": f"{TEST_DIRECTORY}/blueprints/cstar_blueprint_with_yaml_datasets_template.yaml",
+        "template_blueprint_path": f"{TEST_DIRECTORY}/integration_tests/blueprints/cstar_blueprint_with_yaml_datasets_template.yaml",
         "strs_to_replace": {
             "<input_datasets_location>": f"{CSTAR_TEST_DATA_DIRECTORY/'roms_tools_yaml_files'}",
             "<additional_code_location>": f"{CSTAR_TEST_DATA_DIRECTORY}",
